split lexer file
