+++ include/minishell.h
@@ -6,7 +6,7 @@
 /*   By: yumaohno <yumaohno@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/03/24 18:36:52 by yumaohno          #+#    #+#             */
-/*   Updated: 2023/03/29 03:29:42 by yumaohno         ###   ########.fr       */
+/*   Updated: 2023/03/29 17:04:36 by yumaohno         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -24,6 +24,8 @@
 # include <readline/history.h>
 # include <readline/readline.h>
 # include <sys/wait.h>
+
+# define SINGLE_QUOTE_CHAR '\''
 
 typedef enum e_token_kind
 {
@@ -47,5 +49,11 @@ char	**token_list_to_argv(t_token *tok);
 void	fatal_error(const char *str);
 void	assert_error(const char *str);
 t_token	*create_new_token_list(char *word, t_token_kind kind);
+bool	is_word(char *line);
+t_token	*add_word_to_list(char **rest_line, char *line);
+bool	is_operator(char *line);
+t_token	*add_operator_to_list(char **rest_line, char *line);
+bool	is_blank(char c);
+void	skip_blank(char **skip_line, char *line);
 
 #endif
