+++ src/lexer/token_list_to_argv.c
@@ -1,0 +1,35 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   token_list_to_argv.c                               :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: yumaohno <yumaohno@student.42.fr>          +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2023/03/29 17:07:45 by yumaohno          #+#    #+#             */
+/*   Updated: 2023/03/29 17:07:54 by yumaohno         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+#include "minishell.h"
+
+static char	**tail_recursive(t_token *tok, int nargs, char **argv)
+{
+	if (tok == NULL || tok->kind == TK_EOF)
+		return (argv);
+	argv = realloc(argv, (nargs + 2) * sizeof(char *));
+	argv[nargs] = strdup(tok->word);
+	if (argv[nargs] == NULL)
+		fatal_error("strdup");
+	argv[nargs + 1] = NULL;
+	return (tail_recursive(tok->next, nargs + 1, argv));
+}
+
+char	**token_list_to_argv(t_token *tok)
+{
+	char	**argv;
+
+	argv = calloc(1, sizeof(char *));
+	if (argv == NULL)
+		fatal_error("calloc");
+	return (tail_recursive(tok, 0, argv));
+}
