+++ test/test_operator.c
